Bit Manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int result;
+	unsigned long int result;
 	int num = 0;
 
 	result = n ^ m;
